\name{anomalies}
\alias{anomalies}
\title{Annual Precipitation Anomalies in the U.S.}
\usage{data(anomalies)}
\description{
A numerical matrix of dimension \eqn{7252 \times 3} containing longitude, latitude, and yearly total precipitation anomalies registered at 7,352 location sites in the USA.
}
\format{
A numeric matrix with 7,252 rows and 3 columns:
\describe{
  \item{Column 1}{Longitude}
  \item{Column 2}{Latitude}
  \item{Column 3}{Annual precipitation anomaly}
}
}
\source{
Kaufman, C.G., Schervish, M.J., Nychka, D.W. (2008). Covariance tapering for likelihood-based estimation in large spatial data sets. \emph{Journal of the American Statistical Association, Theory & Methods}, \bold{103}, 1545--1555.
}
\keyword{datasets}
