\name{austemp}
\alias{austemp}
\title{Maximum Australian Temperature}
\usage{data(austemp)}
\description{
A matrix containing maximum temperatures in Australia recorded in July 2011.
}
\format{
A \eqn{446 \times 4}{446 x 4} matrix with the following columns:
\itemize{
  \item Longitude
  \item Latitude
  \item Maximum temperature
  \item Geometric temperature covariate
}
}
\source{
Bevilacqua, M., Caamaño, C., Morales-Oñate, V., Arellano-Valle, R. B. (2020). \cr
Non-Gaussian Geostatistical Modeling using (skew) t Processes. \cr
\emph{Scandinavian Journal of Statistics}.
}
\keyword{datasets}
