\name{CkLikelihood}
\alias{CkLikelihood}
\encoding{UTF-8}
\title{Checking Composite-likelihood Type}
\description{
  Subroutine called by InitParam.
  The procedure controls the type of the composite-likelihood inserted
  by the users.
}
\usage{
CkLikelihood(likelihood)
}
\arguments{
  \item{likelihood}{String; the configuration of the composite
    likelihood. \code{Marginal} is the default.}
}

\seealso{\code{\link{GeoFit}}}



\author{Moreno Bevilacqua, \email{moreno.bevilacqua@uv.cl},\url{https://sites.google.com/a/uv.cl/moreno-bevilacqua/home},
Víctor Morales Oñate, \email{victor.morales@uv.cl}, \url{https://sites.google.com/site/moralesonatevictor/}
}

\keyword{Composite}
