\name{CheckSph}
\alias{CheckSph}
\encoding{UTF-8}
\title{Checking if a covariance is valid only on the sphere}
\description{
  Subroutine called by InitParam.
  The procedure controls if a covariance model is valid only on the sphere.
}
\usage{
CheckSph(numbermodel)}
\arguments{
   \item{numbermodel}{Numeric; the code number for the covariance model.}
}

\details{The function checks if a covariance is valid only on the sphere}
\value{Returns TRUE or FALSE} 


\author{Moreno Bevilacqua, \email{moreno.bevilacqua89@gmail.com},\url{https://sites.google.com/view/moreno-bevilacqua/home},
Víctor Morales Oñate, \email{victor.morales@uv.cl}, \url{https://sites.google.com/site/moralesonatevictor/},
Christian", Caamaño-Carrillo, \email{chcaaman@ubiobio.cl},\url{https://www.researchgate.net/profile/Christian-Caamano}
}

\keyword{Composite}
