\name{Cubatemp}
\alias{Cubatemp}
\title{January monthly average temperature in Cuba between 1970-2000}
\usage{data(Cubatemp)}
\description{
  A (\eqn{48679x 3})-matrix containing spatial january  average temperature (°C) observed 
   between 1970-2000 in Cuba with associated  UTM coordinates. The data has been retrieved using the   package \pkg{geodata} 
that allows to download climate data from WorldClim version 2.1.
UTM coordinates has been obtained using zone 18 and datum WGS84 and rescaled by 1000 to have distances in KM.
}
\format{A numerical matrix of dimension \eqn{48679 x 3}.}
\source{
 Fick, S.E., Hijmans, R.J. (2017) WorldClim 2: new 1km spatial resolution climate surfaces for global land areas.
  \emph{International Journal of Climatology}, \bold{37}, 4302--4315.
}

\keyword{datasets}

