\name{winds.coords}
\alias{winds.coords}
\title{Weather Stations of the Irish Daily Wind Speeds}
\usage{data(irishwinds)}
\description{
  A data frame containing information about the weather stations
  where the data are recorded in Ireland.
}
\format{A data frame containing site - the name of the city (character),
  abbr - the abbrevation (character), elev - the elevation (numeric),
  lat - latitude (numeric) and lon  - longitude.}
\source{
  Haslett, J. and Raftery, A. E. (1989),
  Space-time modelling with long-memory dependence: assessing
  Ireland's wind-power resource (with discussion),
  \emph{Applied Statistics}, 38, 1--50.
}
\keyword{datasets}
