\name{winds}
\alias{winds}
\title{Irish Daily Wind Speeds}
\usage{data(irishwinds)}
\description{
  A matrix containing daily wind speeds, in
  kilometers per hour, from 1961 to 1978 at 12 sites in Ireland.
}
\format{A (\eqn{6574 \times 11}{6574 x 11})-matrix containing wind speed observations.}
\source{
  Haslett, J. and Raftery, A. E. (1989),
  Space-time modelling with long-memory dependence: assessing
  Ireland's wind-power resource (with discussion),
  \emph{Applied Statistics}, 38, 1--50.
}
\keyword{datasets}
