\name{Jamaicatemp}
\alias{Jamaicatemp}
\title{December monthly average temperature in Jamaica between 1970-2000}
\usage{data(Jamaicatemp)}
\description{
  A (\eqn{13530x 3})-matrix containing spatial december  average temperature (°C) observed 
   between 1970-2000 in Jamaica with associated  UTM coordinates. The data has been retrieved using the   package \pkg{geodata} 
that allows to download climate data from WorldClim version 2.1.
UTM coordinates has been obtained using zone 18 and datum WGS84 and rescaled by 1000 to have distances in KM.
}
\format{A numerical matrix of dimension \eqn{13530 x 3}.}
\source{
 Fick, S.E., Hijmans, R.J. (2017) WorldClim 2: new 1km spatial resolution climate surfaces for global land areas.
  \emph{International Journal of Climatology}, \bold{37}, 4302--4315.
}

\keyword{datasets}

