\name{madagascarph}
\alias{madagascarph}
\title{Soil ph of Madagascar}
\usage{data(madagascarph)}
\description{
  A (\eqn{300649x 3})-matrix containing lon/lat  coordinates
and  soil ph in H20  at 300649 location sites in Madagascar at one meter depth over the period 2008–2014. Data obtained using the Geodata package with the function \code{soil_af}. For more info, see https://www.isric.org/projects/soil-property-maps-africa-250-m-resolution
}
\format{A numerical matrix of dimension \eqn{300649 x 3}.}
\source{
Hengl T, Heuvelink GBM, Kempen B, Leenaars JGB, Walsh MG, Shepherd KD, et al. (2015) Mapping Soil Properties of Africa at 250 m Resolution: Random Forests Significantly Improve Current Predictions. PLoS ONE 10(6): e0125814. doi:10.1371/journal.pone.0125814
}

\keyword{datasets}

