\name{madagascartemp}
\alias{madagascartemp}
\title{July Average Temperature of Madagascar}
\usage{data(madagascartemp)}
\description{
A \eqn{2500 \times 3}{2500 x 3} matrix containing UTM coordinates and July average temperatures at 2500 location sites in Madagascar, averaged over the period 1970--2000.\\
Data obtained using the \pkg{Geodata} package with the function \code{worldclim_country}.
}
\format{
A numerical matrix of dimension \eqn{2500 \times 3}{2500 x 3}.
}
\source{
Fick, S.E. and Hijmans, R.J. (2017).\\
WorldClim 2: new 1 km spatial resolution climate surfaces for global land areas.\\
\emph{International Journal of Climatology}, \bold{37}(12), 4302--4315.
}
\keyword{datasets}