\name{MatDecomp}
\alias{MatDecomp}
\encoding{UTF-8}
\title{Matrix decomposition}
\description{
  Matrix decomposition.
}
\usage{
MatDecomp(mtx, method)
}
\arguments{
  \item{mtx}{numeric; a square positive or semipositive definite matrix.}
  \item{method}{string; the type of matrix decomposition. 
          Two possible choices: \code{cholesky} and \code{svd}.}
}

\value{Return a matrix decomposition}

\details{ Decomposition of a square positive or positive semidefinite matrix.} 

\author{Moreno Bevilacqua, \email{moreno.bevilacqua89@gmail.com},\url{https://sites.google.com/view/moreno-bevilacqua/home},
Víctor Morales Oñate, \email{victor.morales@uv.cl}, \url{https://sites.google.com/site/moralesonatevictor/},
Christian", Caamaño-Carrillo, \email{chcaaman@ubiobio.cl},\url{https://www.researchgate.net/profile/Christian-Caamano}
}



\keyword{Composite}
