\name{spanish_wind}
\alias{spanish_wind}
\title{August monthly average wind speed in Spain between 1970-2000}
\usage{data(spanish_wind)}
\description{
  A (\eqn{6000x 3})-matrix containing lon/lat 
and  august monthly average wind speed (2 m above the ground, meter/second) registered at 6000 location sites in the Iberian peninsula.
Data obtained from WorldClim version 2.1}
\format{A numerical matrix of dimension \eqn{6000 x 3}.}
\source{
 Fick, S.E., Hijmans, R.J. (2017) WorldClim 2: new 1km spatial resolution climate surfaces for global land areas.
  \emph{International Journal of Climatology}, \bold{37}, 4302--4315.
}

\keyword{datasets}

