\name{anomalies}
\alias{anomalies}
\title{Annual precipitation anomalies in U.S.}
\usage{data(anomalies)}
\description{

  A (\eqn{7252 x 3})-matrix containing lon/lat 
and yearly total precipitation anomalies registered at 7.352 location sites in USA.
For more details see   \url{http://www.image.ucar. edu/Data/precip_tapering/}.
}
\format{A numerical matrix of dimension \eqn{7252 x 3}.}
\source{
  Kaufman, C.G., Schervish, M.J., Nychka, D.W. (2008)
  Covariance tapering for likelihood-based estimation in large spatial data sets.
  \emph{Journal of the American Statistical Association, Theory & Methods}, \bold{103}, 1545--1555.
}
\keyword{datasets}
