\name{austemp}
\alias{austemp}
\title{Maximum australian temperature}
\usage{data(austemp)}
\description{
  A matrix containing maximum  temperature in Australia in July 2011.
}
\format{A (\eqn{446 \times 4}{446 x 4})-matrix containing longitude, latitude, maximum temperature,
 and the 'so called'  geometric temperature covariate.}
\source{
  Bevilacqua M., Caamaño C., Morales-Oñate V.,   Arellano-Valle R. B. (2020)
  Non-Gaussian Geostatistical Modeling using (skew) t Processes,
  \emph{Scandinavian Journal of  Statistics}.
}
\keyword{datasets}
