\name{CheckSph}
\alias{CheckSph}
\encoding{UTF-8}
\title{Checking if a covariance is valid only on sphere}
\description{
  Subroutine called by InitParam.
  The procedure controls if a covariance model is valid only on the sphere.
}
\usage{
CheckSph(numbermodel)}
\arguments{
   \item{numbermodel}{Numeric; the code number for the covariance model.}
}

\value{Returns TRUE or FALSE} 


\author{Moreno Bevilacqua, \email{moreno.bevilacqua@uv.cl},\url{https://sites.google.com/a/uv.cl/moreno-bevilacqua/home},
Víctor Morales Oñate, \email{victor.morales@uv.cl}, \url{https://sites.google.com/site/moralesonatevictor/}
}

\keyword{Composite}
