\name{CkCorrModel}
\alias{CkCorrModel}
\encoding{UTF-8}
\title{Checking Correlation Model}
\description{
  The procedure controls if the correlation model
  inserted is correct.
}
\usage{
CkCorrModel(corrmodel)
}
\arguments{
  \item{corrmodel}{String; the name of a correlation model, for the
    description see \code{\link{GeoCovmatrix}}.}
}

\value{Return a number associated to a given correlation model if the model
is considered in the package. Otherwise return NULL.} 




\author{Moreno Bevilacqua, \email{moreno.bevilacqua@uv.cl},\url{https://sites.google.com/a/uv.cl/moreno-bevilacqua/home},
Víctor Morales Oñate, \email{victor.morales@uv.cl}, \url{https://sites.google.com/site/moralesonatevictor/}
}

\keyword{Composite}
