\name{CkModel}
\alias{CkModel}
\encoding{UTF-8}
\title{Checking Random Field type}
\description{
  Subroutine called by InitParam.
  The procedure controls the type of random field inserted by the users.
}
\usage{
CkModel(model)
}
\arguments{
   \item{model}{String; the density associated to the likelihood
    objects. \code{Gaussian} is the default.}
}

\seealso{\code{\link{GeoFit}}}


\author{Moreno Bevilacqua, \email{moreno.bevilacqua@uv.cl},\url{https://sites.google.com/a/uv.cl/moreno-bevilacqua/home},
Víctor Morales-Oñate, \email{victor.morales@uv.cl}
}

\keyword{Composite}
