\name{CorrelationPar}
\alias{CorrelationPar}
\encoding{UTF-8}
\title{Lists the Parameters of a Correlation Model}
\description{
  Subroutine called by InitParam and other procedures.
  The procedure returns a list with the parameters of a given
  correlation model.
}
\usage{
CorrelationPar(corrmodel)
}
\arguments{
  \item{corrmodel}{Integer; an integer associated to a given correlation model.}
}

\value{Return a vector string of correlation parameters.}


\details{The function return a list with the Parameters of a Correlation Model}

\seealso{\code{\link{GeoFit}}}



\author{Moreno Bevilacqua, \email{moreno.bevilacqua89@gmail.com},\url{https://sites.google.com/view/moreno-bevilacqua/home},
Víctor Morales Oñate, \email{victor.morales@uv.cl}, \url{https://sites.google.com/site/moralesonatevictor/},
Christian", Caamaño-Carrillo, \email{chcaaman@ubiobio.cl},\url{https://www.researchgate.net/profile/Christian-Caamano}
}


\keyword{Composite}
