\name{madagascartemp}
\alias{madagascartemp}
\title{July average  temperature of Madagascar}
\usage{data(madagascartemp)}
\description{
  A (\eqn{2500x 3})-matrix containing UTM  coordinates and  July average temperature at 2500 location sites in Madagascar  over the period 1970–2000. Data obtained using the Geodata package with the function \code{worldclim_country}.
}
\format{A numerical matrix of dimension \eqn{2500 x 3}.}
\source{
Fick, S.E. and R.J. Hijmans, 2017. WorldClim 2: new 1km spatial resolution climate surfaces for global land areas. International Journal of Climatology 37 (12): 4302-4315.
}

\keyword{datasets}

