\name{MatDecomp}
\alias{MatDecomp}
\encoding{UTF-8}
\title{Matrix decomposition}
\description{
  Matrix decomposition.
}
\usage{
MatDecomp(mtx, method)
}
\arguments{
  \item{mtx}{numeric; a square positive or semipositive definite matrix.}
  \item{method}{string; the type of matrix decomposition. 
          Two possible choices: \code{cholesky} and \code{svd}.}
}

\author{Moreno Bevilacqua, \email{moreno.bevilacqua@uv.cl},\url{https://sites.google.com/a/uv.cl/moreno-bevilacqua/home},
Víctor Morales Oñate, \email{victor.morales@uv.cl}
}



\keyword{Composite}
